+++ src/main/java/chess/ConsoleApplication.java
@@ -1,8 +1,10 @@
 package chess;
 
+import chess.console.ConsoleGameRunner;
+
 public class ConsoleApplication {
 
     public static void main(String[] args) {
-        new ConsoleController().run();
+        new ConsoleGameRunner().run();
     }
 }
