+++ src/main/java/chess/domain/command/Command.java
@@ -2,6 +2,7 @@ package chess.domain.command;
 
 import chess.domain.board.Board;
 import chess.domain.board.Position;
+import chess.domain.board.StatusResult;
 import chess.domain.piece.Piece;
 import chess.domain.state.State;
 import java.util.Map;
