+++ src/main/java/chess/view/InputView.java
@@ -6,11 +6,8 @@ public class InputView {
 
     private static final Scanner scanner = new Scanner(System.in);
 
-    public static String requestStartOrEnd() {
+    public static String requestToInputCommand() {
         return scanner.nextLine();
     }
 
-    public static String requestCommand() {
-        return scanner.nextLine();
-    }
 }
